# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: MediVault API
  version: 0.21.0
  description: |
    Patient-centric medical records platform. Patients own their data, doctors
    get scoped time-bound access, caregivers act on behalf of linked patients.

    **Authentication.** All endpoints except `/auth/*` and `/emergency/:patientCode`
    require a Bearer JWT in the `Authorization` header. Tokens are issued at
    `/auth/login` and `/auth/verify-otp`; refresh tokens rotate at `/auth/refresh`.

    **Roles.** `PATIENT`, `DOCTOR`, `CAREGIVER`, `ADMIN`. Endpoints declare which
    roles can call them via the role-gate preHandler. Mismatched role returns
    `403 forbidden`.

    **Error shape.** All errors return `{ "error": "<code>", "message": "<human>"}`.
    Validation errors additionally include `issues: [{path, message, code}]`.

  contact:
    name: MediVault
    url: https://medivault.app
servers:
  - url: http://localhost:3001/v1
    description: Local development
  - url: https://api.medivault.app/v1
    description: Production (not deployed yet)
...
paths:
...
  /records/presign:
    post:
      tags: [Records]
      summary: Get a pre-signed S3 PUT URL
      description: File never passes through the API. Client PUTs directly to S3 with the returned URL.
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mimeType, sizeBytes, sha256]
              properties:
                mimeType: { type: string, enum: ['application/pdf', 'image/jpeg', 'image/png'] }
                sizeBytes: { type: integer, maximum: 26214400, description: '25 MB cap' }
                sha256: { type: string, pattern: '^[a-f0-9]{64}$' }
      responses:
        '200':
          description: Pre-signed URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  uploadUrl: { type: string, format: uri }
                  s3Key: { type: string }
                  expiresInSeconds: { type: integer }
  /records/confirm:
    post:
      tags: [Records]
      summary: Confirm a completed S3 upload — creates the record + queues AI extraction
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConfirmRecordBody' }
      responses:
        '201':
          description: Record created
          content:
            application/json:
              schema: { type: object, properties: { recordId: { type: string, format: uuid } } }
        '403': { description: s3Key not for the calling patient, $ref: '#/components/responses/Error' }
  /records:
    get:
      tags: [Records]
      summary: List own records
      security: [{ bearer: [] }]
      parameters:
        - { in: query, name: category, schema: { $ref: '#/components/schemas/RecordCategory' } }
        - { in: query, name: limit, schema: { type: integer, default: 50, maximum: 100 } }
        - { in: query, name: cursor, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Records page }
  /records/{id}:
    get:
      tags: [Records]
      summary: Get a single record (with extracted FHIR)
      security: [{ bearer: [] }]
      parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
      responses:
        '200': { description: Record }
        '404': { description: Not found, $ref: '#/components/responses/Error' }
    delete:
      tags: [Records]
      summary: Soft-delete a record
      security: [{ bearer: [] }]
      parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
      responses:
        '204': { description: Deleted }
  /records/{id}/view-url:
    get:
      tags: [Records]
      summary: Get a short-lived signed GET URL for the record's file
      security: [{ bearer: [] }]
      parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
      responses:
        '200':
          description: Signed URL
          content:
            application/json:
              schema:
                type: object
                properties: { url: { type: string, format: uri }, expiresInSeconds: { type: integer } }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RecordCategory:
      type: string
      enum: [LAB_RESULT, PRESCRIPTION, IMAGING, DISCHARGE_SUMMARY, CONSULTATION_NOTE, VACCINATION, INSURANCE, OTHER]
    ConfirmRecordBody:
      type: object
      required: [s3Key, category, title, mimeType, sizeBytes, sha256]
      properties:
        s3Key: { type: string }
        category: { $ref: '#/components/schemas/RecordCategory' }
        title: { type: string, minLength: 1, maxLength: 200 }
        notes: { type: string, maxLength: 2000 }
        recordedAt: { type: string, format: date-time }
        mimeType: { type: string, enum: ['application/pdf', 'image/jpeg', 'image/png'] }
        sizeBytes: { type: integer, maximum: 26214400 }
        sha256: { type: string, pattern: '^[a-f0-9]{64}$' }
...
    Error:
      type: object
      properties:
        error: { type: string, description: 'machine-readable code, e.g. forbidden, validation_error' }
        message: { type: string }
        issues:
          type: array
          description: 'Present on validation errors only'
          items:
            type: object
            properties:
              path: { type: array, items: { type: string } }
              message: { type: string }
              code: { type: string }
  responses:
...
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }